disable README tests, moved pem mod into schannel
